Render provider output
