+++ cloud/common/deploy/output/noninteractive/noninteractive.go
@@ -1,0 +1,42 @@
+// Copyright 2021 Nitric Technologies Pty Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package noninteractive
+
+import (
+	"io"
+	"strings"
+
+	"github.com/charmbracelet/log"
+)
+
+type NonInteractiveOutput struct {
+	logger *log.Logger
+}
+
+func (n *NonInteractiveOutput) Write(bytes []byte) (int, error) {
+	msg := string(bytes)
+	msg = strings.TrimSuffix(msg, "\n")
+	n.logger.Print(msg)
+	return len(bytes), nil
+}
+
+func NewNonInterativeOutput(output io.Writer) io.Writer {
+	logger := log.New(output)
+	logger.SetPrefix("")
+
+	return &NonInteractiveOutput{
+		logger: logger,
+	}
+}
